C++ A Speed Breaker
Write a program to display the floating number with a given precision value.

Input:
The first line of input contains integer T denoting the number of test cases. For each test case, we take two inputs, one floating number a and an integer b for the precision of that floating number. 

Output:
For each test case we get a single line output i.e the floating number with given precision.

Constraints:
1<=T<=100;
 1<=b<=20;

Example:

Input: 
2
4.567 2
8.92 3
Output:
4.6
8.92